Refactor getMinimumOrderHandlers into smaller handler files.
