+++ js/src/components/shipping-rate-section/minimum-order-card/handlers/index.js
@@ -1,0 +1,3 @@
+export { default as getMinimumOrderAddHandler } from './getMinimumOrderAddHandler';
+export { default as getMinimumOrderChangeHandler } from './getMinimumOrderChangeHandler';
+export { default as getMinimumOrderDeleteHandler } from './getMinimumOrderDeleteHandler';
